Restored hotspot functionality. Restored SSH functionality.
update_networking.sh changes:
Added `check_hotspotstate` to main loop.
Added a check in `check_wifi` to test hotspot state.

Info: When WiFi is started at the same time as the hotspot, wpa_supplicant is quicker to grab wlan0 which results in hotspot failing to come up

Added dropbear to .tmp_update/bin from this PR:https://github.com/OnionUI/Onion/pull/790 to restore SSH functionality and break the SSHState loop pending review of PR.

+++ static/build/.tmp_update/script/update_networking.sh
@@ -13,31 +13,36 @@ main() {
     check_telnetstate 
     check_ntpstate 
     check_httpstate
+	check_hotspotstate
 }
 
 
 check_wifi() {
-    if wifi_enabled; then
-        if ! ifconfig wlan0 || [ -f /tmp/restart_wifi ]; then
-            if [ -f /tmp/restart_wifi ]; then
-                pkill -9 wpa_supplicant
-                pkill -9 udhcpc
-                rm /tmp/restart_wifi
-            fi
+	if flag_enabled HotspotState; then
+		return
+	else
+		if wifi_enabled; then
+			if ! ifconfig wlan0 || [ -f /tmp/restart_wifi ]; then
+				if [ -f /tmp/restart_wifi ]; then
+					pkill -9 wpa_supplicant
+					pkill -9 udhcpc
+					rm /tmp/restart_wifi
+				fi
 
-            log "Network Checker: Initializing wifi..."
+				log "Network Checker: Initializing wifi..."
 
-            /customer/app/axp_test wifion
-            sleep 2 
-            ifconfig wlan0 up
-            $miyoodir/app/wpa_supplicant -B -D nl80211 -iwlan0 -c /appconfigs/wpa_supplicant.conf
-            udhcpc -i wlan0 -s /etc/init.d/udhcpc.script &
-        fi
-    else
-        pkill -9 wpa_supplicant
-        pkill -9 udhcpc
-        /customer/app/axp_test wifioff
-    fi
+				/customer/app/axp_test wifion
+				sleep 2 
+				ifconfig wlan0 up
+				$miyoodir/app/wpa_supplicant -B -D nl80211 -iwlan0 -c /appconfigs/wpa_supplicant.conf
+				udhcpc -i wlan0 -s /etc/init.d/udhcpc.script &
+			fi
+		else
+			pkill -9 wpa_supplicant
+			pkill -9 udhcpc
+			/customer/app/axp_test wifioff
+		fi
+	fi
 }
